TUDOIGUAL/requirements.txt
# Add your requirements here like:
pytest
#pytest-cov
pycrypto
pyscaffold